invest/requirements.txt
dotenv
python-dotenv
langchain_openai
pymongo
sqlalchemy
pandas
numpy
psycopg2-binary